tests/requirements.txt
setuptools
flake8
seleniumbase
pytest-soft-assertions

